refactor(mcp): isolate http routes from lifecycle wiring
